refactor: enums moved to their own package
