+++ backend/src/main/java/com/rokaly/sgm/dto/MachineDTO.java
@@ -1,6 +1,6 @@
 package com.rokaly.sgm.dto;
 
-import com.rokaly.sgm.model.Status;
+import com.rokaly.sgm.utils.enums.Status;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
