+++ backend/src/main/java/com/rokaly/sgm/model/Machine.java
@@ -2,6 +2,7 @@ package com.rokaly.sgm.model;
 
 import com.rokaly.sgm.dto.MachineDTO;
 import com.rokaly.sgm.dto.PutMachineDTO;
+import com.rokaly.sgm.utils.enums.Status;
 import jakarta.persistence.*;
 
 import java.util.List;
